Major changes to support
(A) inlining of the member functions of merge management objects, and

(B) in the case of a user-specified comparison object, inlining of the
member function "compare" of that object.

Part (A) affects the performance of all three polymorphs of AMI_sort
(comparison operator, comparison function, and comparison class
versions). It requires that the base classes for merge management
objects not contain virtual definitions for those member functions that we
want to be inlined. This caused major changes in this module
(ami_sort_single.h).

Part (B) affects the comparison class version of AMI_sort. Since the
user-specified comparison object is used by pqueue_heap_obj (in
pqueue_heap.h) for merging multiple runs (streams), changes were
required to pqueue_heap.h to ensure that this worked. In addition,
main memory sorting (see quicksort.h) also requires many accesses of
the user-specified comparison object. This is used primarily in
setting up the initial sorted runs prior to the merge. Changes were
needed to quicksort.h, therefore, to support part (B).