Assorted const fixes. This removes a few hundred compilation warnings when compiling libtpie and the tests on  gcc version 4.3.2 (Ubuntu 4.3.2-1ubuntu11). TPIE now compiles warning free (save for a single warning in one test) on that compiler!
Fix a very strange way of doing pq_merge_heap::top. It looks like a way of ensuring the returned reference is valid even after push/pop operations that might alter heap[0]. But that is a red herring, it's never safe to assume that references stay valid across operations and it broke upen another invocation of "top" anyways. Besides, _copying_ the element into a temporary variable and then returning a const reference to this variable defeats the entire point of returning const references (saving the copy).