Get rid of file::update_vars(). All it did was keep file_base::stream::m_block->size and m_file.m_size up to date, but this is already done in write_update.
Get rid of multiplication in file_base::stream::write_update(). This method is executed for each item written, so caching the multiplication is beneficial.

Add an integer max without branching. This speeds up writes on Windows, but slows it down on Linux.

Get rid of try/catch in the AMI stream compat layer. try/catch without throw is not zero-cost on Windows.