Rework pipelining merge sort parameter calculation.
Now, you supply distinct memory limits for each of the merge sort
phases, and the code has been reworked to account for differences in
memory limits. Previously, sort_parameters just kept one fanout and one
run length. Now, we also note if the item threshold for internal
reporting is less than the run length, and if the fanout in the final
merge phase is less than the fanout in the merge tree phase.