Check if pipeline relation graph is connected.
When a pipeline_impl instance is created using the pipe_base classes, a
recursive descent through the factories check that the segment_map
instance referenced in each pipe_segment is the same throughout the
pipeline. This catches the common implementation mistake of missing an
add_push_destination/add_pull_destination/add_dependency in the
constructor of the pipe_segment implementation.