Remove parallel() restriction on item_type.
Previously, it was a requirement that the item_type of the pipeline to
parallelize did not depend on the destination type supplied to the
pipeline. Initial refactoring of the parallel.h code made this
requirement superfluous, since the node that the pipeline to parallelize
pushes to (pipelining::parallel_bits::after) no longer depended on the
parallel state type. However, the restriction remained in place.