Enable support for the C++11 override keyword. (#54)
Adds cpp0xtest/override_keyword.cpp which is an example program using
the override keyword.

In CMakeLists.txt, override_keyword.cpp is attempted compiled, and on
success, TPIE_CPP_OVERRIDE_KEYWORD is defined.

In config.h.cmake, a #define macro is inserted to turn `override` into
the empty string when it is not supported by the compiler.