GetGitRevisionDescription: Set GIT_PARENT_DIR to CMAKE_CURRENT_SOURCE_DIR.
When TPIE is used as a submodule, CMAKE_SOURCE_DIR points to the source
dir of the parent repository instead of a directory within TPIE. This
means we report the Git information of the parent repository instead of
TPIE, which is a bit misleading.

CMAKE_CURRENT_SOURCE_DIR points to "the directory where the currently
processed CMakeLists.txt is located in", meaning a parent directory of
this will contain the .git entry of TPIE we are looking for.