unittest.h: Make test_log_target flush its ostream in log().
This might lead to a performance hit in unit tests, but production code
is not affected at all since it does not use unittest.h.

log() is called whenever the std::ostream buffer of the TPIE log is
flushed. The progress_indicator_arrow calls std::flush on its
std::ostream whenever it has been refreshed. If we do not flush in
test_log_target::log(), the progress bar only rarely gets updated, since
the std::ostream buffer is much larger than the size of a single
progress_indicator_arrow update.