Introduce node::propagate and change node::begin order.
node::propagate is called in a topological order of the item flow graph,
which was previously also the case for node::begin. It is permitted to
call fetch and forward in propagate, but not push and pull.

We change node::begin to be called in reverse topo order of the actor
graph, such that it is permitted to call push and pull in begin.

This will break existing code that relies on fetching and forwarding in
begin.