phase::go: If no initiator has been found, throw no_initiator_node.
Previously, there was no check, and a null pointer dereference would
crash the program with a segfault.

A pipelining phase can only have no initiator when there is a cycle in
the actor graph, so this should not happen in practice. However, it is
relatively easy to set up a program that constructs a pipelining graph
with a cycle in the actor graph, and the framework should still handle
that gracefully.