serialization_sort.h: Handle empty input as a special case.
If no input is pushed, set `m_reportInternal` to true and
`m_nextInternalItem` to the null pointer.

When `merge_runs` is then called, nothing is done since
`m_reportInternal` is true.

When `can_pull` is then called, it returns false since
`m_reportInternal` is true and `m_nextInternalItem` is the null pointer.