pipeline::fetch: Do not pass a temporary to boost::any_cast
boost::any_cast<T>(const any<T> &) coerces the payload to be const,
which is invalid if the wrapped object was not originally const.

This happens e.g. when boost::any contains an int which is wrapped in a
boost::any::holder<int>;
boost::any_cast<int>(const any<int> &) will try to coerce this into a
boost::any::holder<const int>
which results in undefined behavior.