Allow using unique_ptrs as pipeline datastructunes, and make the datastructure go away at the right time.
Datastructure entries are now stored as any_uncopyiable, such that a copy constructor is not needed.
Whenever a pipeline phase ends, all datastructures, whos last usage is that phase, are removed from the datastructure hashmap, meaning
that in the case of unique ptrs they are freed, and in the case of shared ptrs they are decrefed.