Use /var/tmp instead of /tmp as default when /tmp is tmpfs
This is for linux only:
If the temporary directory is not specified in an env var, we want to default to /tmp like boost does.
However on some linux systems /tmp is a tmpfs filesystem meaning it is stored in RAM.
If this is the case we want to use /var/tmp as the default instead.